add JWT authentication

+++ controllers/api/v1/post_api.js
@@ -26,7 +26,7 @@ module.exports.index = async function(request, response) {
 module.exports.destroy = async function(request, response) {
     try {
         let post = await Post.findById(request.params.id);
-        // if (post.user == request.user.id) {
+        if (post.user == request.user.id) {
             post.remove();
 
             await Comment.deleteMany({
@@ -36,10 +36,11 @@ module.exports.destroy = async function(request, response) {
             return response.json(200, {
                 message: 'Post and associated comments deleted successfully!'
             });
-        // } else {
-        //     request.flash('error', error);
-        //     return response.redirect('back');
-        // }
+        } else {
+            return response.json(401, {
+                message: 'You cannot delete this post!'
+            })
+        }
     } catch(error) {
         console.log("Error ", error);
         return response.json(500, {
